+++ CHANGELOG.md
@@ -27,8 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#280](https://github.com/CCSDForge/episciences/issues/280): set spellcheck to false in dynamic datatable search boxes.
 - [#281](https://github.com/CCSDForge/episciences/issues/281): Submit an article > Guidelines section: harmonization of journal/review terms [FR]
 - Now, when an article is accepted, all unanswered invitations are deleted. The reviewer is informed of this action.
+- Add ISSN pending until the ISSN is actually issued
 
 ### Fixed
+- Fix incorrect HTML in footer
 - Incorrect paper status LMCS #10145 (related to https://github.com/CCSDForge/episciences/blob/main/CHANGELOG.md#10273---2022-03-18)
 - Empty %%REQUESTER_EXPRESSION%% tag (related to [#187](https://github.com/CCSDForge/episciences/issues/187)
 - Incorrect status jdmdh#10203 (related with git#372)